Move ObjectSet and RefCounted to Helpers
Consolidate all "generally useful, stable APIs" in Helpers, and
also expose RefCounted publicly, so thingengine-core can use it.
