+++ lib/helpers/index.js
@@ -16,4 +16,6 @@ module.exports = {
     PollingStream: require('./polling'),
     Rss: require('./rss'),
     Xml: require('./xml'),
+    RefCounted: require('./ref_counted'),
+    ObjectSet: require('./object_set'),
 };
