+++ lib/helpers/object_set.js
