+++ lib/helpers/ref_counted.js
