+++ lib/messaging.js
@@ -11,7 +11,7 @@
 
 const events = require('events');
 
-const RefCounted = require('./ref_counted');
+const RefCounted = require('./helpers/ref_counted');
 
 class MessagingFeed extends RefCounted {
     // events: new-message, incoming-message, outgoing-message, members-changed
